Simplify definition of NaN
Replace u64->double bitcast with 0.0/0.0.

Change definition of JS_FLOAT64_NAN in the same way because that lets us
drop <math.h> from quickjs.h.

Replace INFINITY and -INFINITY with 1.0/0.0 and -1.0/0.0 respectively
because we already defined globalThis.INFINITY that way and so we might
as well go all the way.