Add pre-init ICs
Defer creating the inline cache once, to avoid unprofitable busywork
for JS code that only executes once.

It's at best a qualified success. It succeeds at keeping the memory and
CPU footprint of execute-once code down but it slows down the babel
benchmark from the web-tooling-benchmark suite by 500% because that
benchmark is all about property access on tree structures.

Other benchmarks from the suite are unaffected, neither faster or
slower, with the possible exception of the terser benchmark, which
is a few percent slower for the same reason as the babel one, only
much less pronounced.