Fix unnoticed exception in JS_DetectModule
JS_DetectModule calls JS_AddIntrinsicRegExp but that indirectly uses
`ctx->global_obj` when it tries to define `globalThis.RegExp`.

There is no `ctx->global_obj` however because JS_NewContextRaw doesn't
create one. It raised an "not an object" TypeError that went unnoticed
because there is no error checking in JS_NewGlobalCConstructor2.

Using JS_AddIntrinsicRegExpCompiler is better all around because we
only need to parse regexp literals, nothing more.

JS_NewGlobalCConstructor2 should probably be changed to do proper error
checking, and __JS_EvalInternal to return JS_EXCEPTION when an exception
is pending on entry, but I'll save that for another commit.