Add parserless build mode
Add a QJS_DISABLE_PARSER build option that disables the JS source code
parser. The JSON parser is unaffected.

Useful if you want to execute bytecode but not parse source code, e.g.,
for hardening reasons, or due to running in a constrained environment.

Shrinks a release build by about 15%.

Fixes: https://github.com/quickjs-ng/quickjs/issues/976