fixed comments and proper symlink test flag
 const IS_WIN_ADMIN_TEST_FLAG = 0;
modified so that someone can run as admin with this flag on and actually test symlink.
comment above that clarified that "this code is broken".  - That's not my code, my code makes the majority of the std_file finish executing. I just can't readlink from windows , and I can't do that mtime.
...... (but I don't need metric time, I'm not a scientist!