cleaning up and updating
removed the majority of flags. Only left flags that seemed necessary. 
I had a combined js_os_pipe with several #ifdefs inside to separate the OSs. I later restored the non-WIN32 version so that it would not be out of order from the code before. The logic was that a) the export list is maintained in the order of the function calls. and b) the code compare was less of a mess to keep that chunk intact.  I realized there were useless #ifdefs and removed them. I did retest both WIN32 and GNU.