Update quickjs-libc.c
CloseHandle after OpenProcess. 
The documentation wasn't so clear, but this works as it should, and I read that it wastes resources not to. Don't close HNDs to FDs, let them get closed by the user. But close HNDs from PIDs.
Besides, this is a kill and an exit return value. It's stopped!