cwd tested from getcwd
I was building the script path, but I needed to test cwd. 
test262 is usually called from the make folder. 
manually I'm going into tests/ and calling from there so that assert.js is there.
now I look and see if 'tests' is the end of the path, if not I insert it, and I use 'cwd: dir,' 
If someone executes test262 from a build folder, I don't know how they get assert.js, but this will then break.