run-test262: handle scandir() results with DT_UNKNOWN d_type
In case the dirent items returned by scandir() have DT_UNKNOWN as
d_type it is not possible to assume they are "non directories", and
it is needed to stat() them to get the actual type.

Hence, tweak the existing bits to account for that:
- create an helper enum for file type to distinguish whether something
  is for sure a file, whether it is for sure a directory, or it is not
  known
- map d_type to the newly created enum
- adapt consider_test_file() to use the newly created enum, and case
  the file type is not known also to stat() the file to get the actual
  type

On Windows the file type is always known, so there is no need to stat()
files.