Improve handling of set-like .size properties
test262 doesn't test it but the spec mandates that negative .size
properties should result in a RangeError being raised, and that is
indeed what V8 and SpiderMonkey do. Now we do, too.

Adjust for another quirk, which is that Infinity is a legal size,
and, more generally, that sizes > Number.MAX_SAFE_INTEGER are not
illegal - which is okay, because the size is not actually used for
anything, because of course it isn't.