Add ref-counted string slices (#1175)
Instead of copying long substring slices, store a reference to and
offset into the parent string.

Gets more profitable as strings get longer. For 64k string slices,
it's about 6.5x faster than copying.

The downside of ref-counted string slices is that they can keep the
parent string alive for longer than it otherwise would have been,
leading to memory usage that is higher than without string slices.
That's why this optimization is only applied to long-ish slices,
currently slices > 1,024 bytes.

Possible future enhancements are slicing only when the substring
is > x% of the parent string, or copying lazily when the slice
string is the only thing referencing the parent string.