Make JS_ToCString handle string slices correctly
JS_ToCStringLen2 returns a pointer to a string that is prefixed by a
JSString struct. It lets JS_FreeCString free the memory through simple
pointer arithmetic.

JS_ToCStringLen2 has an optimization for ASCII-only 8 bit strings
where it returns early but that optimization is not valid for slice
strings. They reference their string data indirectly and said data is
not prefixed by a JSString.

Fixes: https://github.com/quickjs-ng/quickjs/issues/1178