Revert "Add ref-counted string slices (#1175)"
This change seems to have introduced a seemingly impossible bug in a
real-world code base where values flip to JS_NULL for no discernible
reason.

My best guess is something manipulating string data directly, without
being aware of slice strings, but it's impossible to track down where.
The weird thing is that the value flipping to null isn't even a string,
it appears to be an array.

This reverts commit 08db51a73a9ce3948e9e7ac8142689fe4c83d46f.
This reverts commit 62b4eede257c1ab8f82005e4e2042330a0acb323.

Fixes: https://github.com/quickjs-ng/quickjs/issues/1178