Propagate exceptions in Iterator.concat (#1256)
test262 took a 180 degree turn because the test that is fixed by this
commit is the polar opposite of what was tested by the test suite when
I originally added Iterator.concat.

The old behavior was not very efficient, so good for them for changing
their minds, but next time please do so before I spend an hour
implementing it, thank you very much.