Consolidate helper macros (JS_EXTERN, js_force_inline) in one header
Since the macros like JS_EXTERN are used in more than one place, it is
quite reasonable and helpful to declare them in one place.

This commit, in addition to moving the macro definitions into the new
header quickjs-ng-util.h, also plumbs Windows support for the newly
introduced QUICKJS_NG_API and QUICKJS_NG_FORCE_INLINE.

The new header is installed with quickjs.h, since the latter depends on
the former.

Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>